requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
Django==4.2.5
django-filter==23.2
djangorestframework==3.14.0
flake8==6.1.0
Markdown==3.4.4
mccabe==0.7.0
Pillow==10.0.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
